orders imports and removes basesuite
